bamazonCustomer working

+++ bamazon.sql
@@ -5,7 +5,7 @@ CREATE TABLE products (
 	id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
     product_name VARCHAR(50) NOT NULL,
     department_name VARCHAR(50) NOT NULL,
-    price INT NOT NULL,
+    price DECIMAL(6,2) NOT NULL,
     stock_quantity INT NOT NULL
 );
